Algorithms class split into 3 with respect to their purpose
