+++ src/Spectre.Algorithms/Spectre.Algorithms.csproj
@@ -44,7 +44,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Algorithms.cs" />
+    <Compile Include="Methods\Gmm.cs" />
+    <Compile Include="Methods\Preprocessing.cs" />
+    <Compile Include="Methods\Segmentation.cs" />
     <Compile Include="Parameterization\DivikOptions.cs" />
     <Compile Include="Parameterization\Metric.cs" />
     <Compile Include="Results\DivikResult.cs" />
